feat: usernavbar is now called navbar
